fix: server interrupt on restore backup callback request

+++ server/src/controllers/BackupController.ts
@@ -113,12 +113,14 @@ class BackupController {
       const usuariosExcluidos_collection = conexao_mongodb.getBancoDados().collection('log_usuario_excluido');
       const usuariosExcluidos = await usuariosExcluidos_collection.find().toArray();
 
-      // excluir usuarios no banco
-      await AppDataSource.createQueryBuilder()
-        .delete()
-        .from(Usuario)
-        .where("id IN (:...ids)", { ids: usuariosExcluidos.map(item => item.id_usuario) })
-        .execute();
+      if(usuariosExcluidos.length) {
+        // excluir usuarios no banco
+        await AppDataSource.createQueryBuilder()
+          .delete()
+          .from(Usuario)
+          .where("id IN (:...ids)", { ids: usuariosExcluidos.map(item => item.id_usuario) })
+          .execute();
+      }
     } catch (e) {
       return res.status(500).json({ error: "Erro no servidor", errorCode: "500-server-error", details: { ...e } });
     } finally {
